requirements.txt
alembic==1.7.7
asgiref==3.4.1
backports.zoneinfo==0.2.1
bcrypt==4.0.1
blinker==1.5
certifi==2019.3.9
chardet==3.0.4
charset-normalizer
click==8.0.4
colorama==0.4.5
dnspython==2.0.0
Flask==2.0.3
Flask-JWT-Extended==4.4.1
Flask-Mail==0.9.0
Flask-Migrate==4.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flask-profiler==1.8.1
geographiclib==1.52
geopy==2.3.0
greenlet==2.0.1
gunicorn
idna==2.7
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipwhois==1.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.6
MarkupSafe==2.1.1
numpy==1.19.5
pygeoip==0.3.0
PyJWT==2.6.0
python-decouple==3.8
requests==2.27.0
SQLAlchemy==1.4.46
sqlparse==0.4.1
stem==1.1.1
stripe==2.60.0
toml==0.9.6
types-click==7.1.8
types-Flask==1.1.6
types-Jinja2==2.11.9
types-MarkupSafe==1.1.10
types-Werkzeug==1.0.9
typing_extensions==4.1.0
tzdata==2022.7
ua-parser==0.16.1
urllib3==1.25.3
Werkzeug==2.2.2
WTForms==2.0.1
zipp==3.5.1

